refactor(backend): share multipart Postgres fixture contracts
