+++ apps/backend/src/mediaAssets/multipart/completion/completionReconciliation.postgres.integration.ts
@@ -37,29 +37,13 @@ import {
   buildMediaBlobStorageKey,
   buildMediaMultipartUploadStagingStorageKey,
 } from "../../storageKeys";
-
-type MultipartPayload = Readonly<{
-  userId: string;
-  workspaceId: string;
-  sessionId: string;
-  mediaAssetId: string;
-  replicaId: string;
-  lastOperationId: string;
-  sha256: string;
-  stagingStorageKey: string;
-  blobStorageKey: string;
-  s3UploadId: string;
-  mimeType: string;
-  sizeBytes: number;
-  partSizeBytes: number;
-  partCount: number;
-  sourceUrl: string | null;
-  assetCreatedAt: string;
-  clientUpdatedAt: string;
-  sessionExpiresAt: string;
-  normalizationVersion: string;
-  partsFingerprint: string;
-}>;
+import {
+  createMultipartPayloadFixture,
+  type MultipartPayload,
+  multipartPayloadCompositeRow as multipartRow,
+  multipartPayloadValues as payloadValues,
+  withMultipartPostgresTransaction,
+} from "../postgresTestSupport";
 
 type BeginRow = Readonly<{
   attempt_status: string;
@@ -75,12 +59,7 @@ type StateRow = Readonly<{
   reservation_state: string;
 }>;
 type CountRow = Readonly<{ count: number }>;
-type SqlValue = string | number | null;
 type QueryExecutor = Pick<pg.PoolClient | pg.Pool, "query">;
-
-const multipartRow = `ROW(
-  $3,$4,$5,$6,$7,$8,$9,$10,$11,$12,$13,$14,$15,$16,$17,$18,$19,$20,$21,$22
-)::content.multipart_media_blob_writer_attempt_payload`;
 
 function digest(): string {
   return createHash("sha256").update(randomUUID()).digest("hex");
@@ -89,32 +68,15 @@ function digest(): string {
 function createPayload(
   fixture: PostgresIntegrationFixture,
 ): MultipartPayload {
-  const sessionId = randomUUID();
-  const mediaAssetId = randomUUID();
-  const sha256 = digest();
-  return {
+  return createMultipartPayloadFixture({
     userId: fixture.userId,
     workspaceId: fixture.workspaceId,
-    sessionId,
-    mediaAssetId,
+    mediaAssetId: randomUUID(),
     replicaId: fixture.replicaId,
-    lastOperationId: randomUUID(),
-    sha256,
-    stagingStorageKey:
-      `media/uploads/workspaces/${fixture.workspaceId}/assets/${mediaAssetId}/sessions/${sessionId}`,
-    blobStorageKey: buildMediaBlobStorageKey(sha256),
-    s3UploadId: `upload-${randomUUID()}`,
-    mimeType: "application/octet-stream",
-    sizeBytes: 42,
-    partSizeBytes: 42,
-    partCount: 1,
-    sourceUrl: null,
     assetCreatedAt: fixture.createdAt,
     clientUpdatedAt: fixture.createdAt,
     sessionExpiresAt: new Date(Date.now() + 3_600_000).toISOString(),
-    normalizationVersion: "passthrough-v1",
-    partsFingerprint: digest(),
-  };
+  });
 }
 
 const migration0099LegacyUserId =
@@ -167,31 +129,6 @@ function createMigration0099LegacyPayload(
   };
 }
 
-function payloadValues(payload: MultipartPayload): Array<SqlValue> {
-  return [
-    payload.userId,
-    payload.workspaceId,
-    payload.sessionId,
-    payload.mediaAssetId,
-    payload.replicaId,
-    payload.lastOperationId,
-    payload.sha256,
-    payload.stagingStorageKey,
-    payload.blobStorageKey,
-    payload.s3UploadId,
-    payload.mimeType,
-    payload.sizeBytes,
-    payload.partSizeBytes,
-    payload.partCount,
-    payload.sourceUrl,
-    payload.assetCreatedAt,
-    payload.clientUpdatedAt,
-    payload.sessionExpiresAt,
-    payload.normalizationVersion,
-    payload.partsFingerprint,
-  ];
-}
-
 async function scoped<Result>(
   fixture: PostgresIntegrationFixture,
   callback: (client: pg.PoolClient) => Promise<Result>,
@@ -210,22 +147,11 @@ async function scopedAs<Result>(
   workspaceId: string,
   callback: (client: pg.PoolClient) => Promise<Result>,
 ): Promise<Result> {
-  const client = await fixture.runtimePool.connect();
-  try {
-    await client.query("BEGIN");
-    await client.query(
-      "SELECT set_config('app.user_id',$1,true),set_config('app.workspace_id',$2,true)",
-      [userId, workspaceId],
-    );
-    const result = await callback(client);
-    await client.query("COMMIT");
-    return result;
-  } catch (error) {
-    await client.query("ROLLBACK");
-    throw error;
-  } finally {
-    client.release();
-  }
+  return withMultipartPostgresTransaction(
+    fixture.runtimePool,
+    { userId, workspaceId },
+    callback,
+  );
 }
 
 async function insertSession(
