+++ apps/backend/src/mediaAssets/multipart/writerLifecycle/writerAttempts.postgres.integration.ts
@@ -6,19 +6,18 @@ import test from "node:test";
 import type pg from "pg";
 import { type PostgresIntegrationFixture, withPostgresIntegrationFixture } from "../../../testSupport/postgresIntegration";
 import { buildMediaBlobStorageKey } from "../../storageKeys";
+import {
+  insertMultipartUploadSession as insertSession,
+  type MultipartPayload,
+  multipartPayloadCompositeRow as multipartRow,
+  multipartPayloadValues as multipartValues,
+  withMultipartPostgresTransaction,
+} from "../postgresTestSupport";
 type DirectPayload = Readonly<{
   userId: string; workspaceId: string; mediaAssetId: string; operationId: string;
   replicaId: string; sha256: string; storageKey: string; mimeType: string;
   sizeBytes: number; normalizationVersion: string; sourceUrl: string | null;
   assetCreatedAt: string; clientUpdatedAt: string;
-}>;
-type MultipartPayload = Readonly<{
-  userId: string; workspaceId: string; sessionId: string; mediaAssetId: string;
-  replicaId: string; lastOperationId: string; sha256: string;
-  stagingStorageKey: string; blobStorageKey: string; s3UploadId: string;
-  mimeType: string; sizeBytes: number; partSizeBytes: number; partCount: number;
-  sourceUrl: string | null; assetCreatedAt: string; clientUpdatedAt: string;
-  sessionExpiresAt: string; normalizationVersion: string; partsFingerprint: string;
 }>;
 type BeginRow = Readonly<{
   attempt_status: string; reservation_token: string | null;
@@ -35,8 +34,6 @@ const migration0097 = readFileSync(resolve(
 ), "utf8");
 const directRow = `ROW($3,$4,$5,$6,$7,$8,$9,$10,$11,$12,$13,$14,$15)
   ::content.direct_media_blob_writer_attempt_payload`;
-const multipartRow = `ROW($3,$4,$5,$6,$7,$8,$9,$10,$11,$12,$13,$14,$15,$16,$17,$18,$19,$20,$21,$22)
-  ::content.multipart_media_blob_writer_attempt_payload`;
 function digest(): string {
   return createHash("sha256").update(randomUUID()).digest("hex");
 }
@@ -75,32 +72,15 @@ function directValues(payload: DirectPayload): ReadonlyArray<SqlValue> {
     payload.sizeBytes, payload.normalizationVersion, payload.sourceUrl,
     payload.assetCreatedAt, payload.clientUpdatedAt];
 }
-function multipartValues(payload: MultipartPayload): ReadonlyArray<SqlValue> {
-  return [payload.userId, payload.workspaceId, payload.sessionId, payload.mediaAssetId,
-    payload.replicaId, payload.lastOperationId, payload.sha256,
-    payload.stagingStorageKey, payload.blobStorageKey, payload.s3UploadId,
-    payload.mimeType, payload.sizeBytes, payload.partSizeBytes, payload.partCount,
-    payload.sourceUrl, payload.assetCreatedAt, payload.clientUpdatedAt,
-    payload.sessionExpiresAt, payload.normalizationVersion, payload.partsFingerprint];
-}
 async function scoped<Result>(
   fixture: PostgresIntegrationFixture,
   callback: (client: pg.PoolClient) => Promise<Result>,
 ): Promise<Result> {
-  const client = await fixture.runtimePool.connect();
-  try {
-    await client.query("BEGIN");
-    await client.query( "SELECT set_config('app.user_id',$1,true),set_config('app.workspace_id',$2,true)", [fixture.userId, fixture.workspaceId],
-    );
-    const result = await callback(client);
-    await client.query("COMMIT");
-    return result;
-  } catch (error) {
-    await client.query("ROLLBACK");
-    throw error;
-  } finally {
-    client.release();
-  }
+  return withMultipartPostgresTransaction(
+    fixture.runtimePool,
+    { userId: fixture.userId, workspaceId: fixture.workspaceId },
+    callback,
+  );
 }
 async function beginDirect(
   fixture: PostgresIntegrationFixture, attemptToken: string,
@@ -119,15 +99,6 @@ async function directStatus(
     `SELECT content.${functionName}($1,$2,${directRow},$16) AS status`,
     [attemptToken, reservationToken, ...directValues(payload), cleanupDelayMs],
   )).rows[0].status);
-}
-async function insertSession(
-  executor: QueryExecutor, payload: MultipartPayload, state: "active" | "completing",
-): Promise<void> {
-  await executor.query(
-    `INSERT INTO content.media_upload_sessions (media_upload_session_id,workspace_id,media_asset_id,media_blob_sha256, staging_storage_key,blob_storage_key,s3_upload_id,mime_type,size_bytes, part_size_bytes,part_count,state,source_url,asset_created_at,client_updated_at, last_modified_by_replica_id,last_operation_id,expires_at)
-     VALUES ($1,$2,$3,$4,$5,$6,$7,$8,$9,$10,$11,$12,$13,$14,$15,$16,$17,$18)`,
-    [payload.sessionId, payload.workspaceId, payload.mediaAssetId, payload.sha256, payload.stagingStorageKey, payload.blobStorageKey, payload.s3UploadId, payload.mimeType, payload.sizeBytes, payload.partSizeBytes, payload.partCount, state, payload.sourceUrl, payload.assetCreatedAt, payload.clientUpdatedAt, payload.replicaId, payload.lastOperationId, payload.sessionExpiresAt],
-  );
 }
 async function beginMultipart(
   fixture: PostgresIntegrationFixture, attemptToken: string,
